+++ doc/release_notes.rst
@@ -1,8 +1,11 @@
 Release Notes
 =============
 
-Upcoming Release
-----------------
+.. Upcoming Release
+.. ----------------
+
+Version 0.0.13
+--------------
 
 **New Features**
 
@@ -12,7 +15,6 @@ Upcoming Release
 **Bug Fixes**
 
 * Filtering out zeros is now an optional feature in the `solve` function. The previous behavior of filtering just before the LP file writing, lead to unexpected errors for constraints with only zero terms.
-
 
 
 Version 0.0.12